Rectangle<int> -> Rect

+++ packages/firebase_ml_vision/android/src/main/java/io/flutter/plugins/firebasemlvision/BarcodeDetector.java
@@ -60,10 +60,10 @@ class BarcodeDetector implements Detector {
 
                   Rect bounds = barcode.getBoundingBox();
                   if (bounds != null) {
-                    barcodeMap.put("left", bounds.left);
-                    barcodeMap.put("top", bounds.top);
-                    barcodeMap.put("width", bounds.width());
-                    barcodeMap.put("height", bounds.height());
+                    barcodeMap.put("left", (double) bounds.left);
+                    barcodeMap.put("top", (double) bounds.top);
+                    barcodeMap.put("width", (double) bounds.width());
+                    barcodeMap.put("height", (double) bounds.height());
                   }
 
                   List<int[]> points = new ArrayList<>();
